Extract inference from //compiler/match/inference to //compiler/inference
