+++ compiler/BUILD
@@ -13,6 +13,7 @@ rust_library(
         "*.rs",
         "match_/**/*.rs",
         "fetch/**/*.rs",
+        "inference/**/*.rs",
         "insert/**/*.rs",
         "delete/**/*.rs",
         "expression/**/*.rs",
