+++ compiler/inference/type_annotations.rs
@@ -9,10 +9,8 @@ use std::{
     sync::Arc,
 };
 
-use answer::{variable::Variable, Type};
+use answer::{Type, variable::Variable};
 use ir::pattern::{constraint::Constraint, Vertex};
-
-use crate::match_::inference::pattern_type_inference::TypeInferenceGraph;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct TypeAnnotations {
@@ -21,13 +19,6 @@ pub struct TypeAnnotations {
 }
 
 impl TypeAnnotations {
-    pub(crate) fn build(inference_graph: TypeInferenceGraph<'_>) -> Self {
-        let mut vertex_annotations = BTreeMap::new();
-        let mut constraint_annotations = HashMap::new();
-        inference_graph.collect_type_annotations(&mut vertex_annotations, &mut constraint_annotations);
-        Self::new(vertex_annotations, constraint_annotations)
-    }
-
     pub fn new(
         variables: BTreeMap<Vertex<Variable>, Arc<BTreeSet<Type>>>,
         constraints: HashMap<Constraint<Variable>, ConstraintTypeAnnotations>,
