+++ compiler/lib.rs
@@ -18,6 +18,7 @@ pub mod insert;
 pub mod match_;
 pub mod modifiers;
 pub mod reduce;
+pub mod inference;
 
 macro_rules! filter_variants {
     ($variant:path : $iterable:expr) => {
