+++ executor/tests/execute_expression.rs
@@ -53,11 +53,11 @@ fn compile_expression_via_match(
             translate_match(&mut translation_context, &HashMapFunctionSignatureIndex::empty(), match_)?.finish();
         let variable_mapping = variable_types
             .keys()
-            .map(|name| ((*name).to_owned(), *translation_context.visible_variables.get(*name).unwrap()))
+            .map(|name| ((*name).to_owned(), translation_context.get_variable(*name).unwrap()))
             .collect::<HashMap<_, _>>();
         let variable_types_mapped = variable_types
             .into_iter()
-            .map(|(name, type_)| (*translation_context.visible_variables.get(name).unwrap(), type_))
+            .map(|(name, type_)| (translation_context.get_variable(name).unwrap(), type_))
             .collect::<HashMap<_, _>>();
 
         let expression_binding = match &block.conjunction().constraints()[0] {
