+++ function/lib.rs
@@ -6,7 +6,7 @@
 
 use std::error::Error;
 
-use compiler::match_::inference::FunctionTypeInferenceError;
+use compiler::inference::FunctionTypeInferenceError;
 use encoding::error::EncodingError;
 use error::typedb_error;
 use ir::program::{FunctionReadError, FunctionRepresentationError};
